+++ src/main/java/zos/shell/service/dsn/concat/ConcatCmd.java
@@ -1,4 +1,4 @@
-package zos.shell.service.dsn.concatenate;
+package zos.shell.service.dsn.concat;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
