+++ src/main/java/zos/shell/service/grep/GrepCmd.java
@@ -3,7 +3,7 @@ package zos.shell.service.grep;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import zos.shell.constants.Constants;
-import zos.shell.service.dsn.concatenate.ConcatCmd;
+import zos.shell.service.dsn.concat.ConcatCmd;
 import zos.shell.service.dsn.download.Download;
 import zos.shell.service.memberlst.MemberLst;
 import zos.shell.utility.ResponseUtil;
